refactor(test): 拆解 test_io_modules 与名不副实的 bootstrap 冒烟测试
- test_io_modules.py -> agent/mcp/test_client.py、agent/test_tool_runtime.py、
  agent/tools/test_filesystem.py
- bootstrap/runtime/test_smoke.py 实为三个不相关模块的完整行为测试，
  按被测模块拆到 bootstrap/test_init_workspace.py（新建）、
  bootstrap/test_channels.py、agent/test_config.py；顺带删除该文件里
  从未被调用的 _toml_value/_dump_toml/_write_config 三个死函数

Refs #194
